# hotel-management-ui/src/openApi/openApi.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "openApi specification - user19",
      "description": "OpenApi documentation for hotel management",
      "termsOfService": "terms of service",
      "contact": {
        "name": "user19",
        "url": "http://user19.com/website",
        "email": "contact@User19.com"
      },
      "license": {
        "name": "licence name",
        "url": "http://some-url.com"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:9090/api/v1",
        "description": "local Env"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
      "/rooms/updateRoomPhoto/{roomId}": {
        "post": {
          "tags": [
            "room"
          ],
          "operationId": "updateRoomPhoto",
          "parameters": [
            {
              "name": "roomId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "required": [
                    "file"
                  ],
                  "type": "object",
                  "properties": {
                    "file": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "/rooms/addNewRoom": {
        "post": {
          "tags": [
            "room"
          ],
          "operationId": "addNewRoom",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomRegistrationRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/booking/book-room/{roomId}/{userId}": {
        "post": {
          "tags": [
            "Booking "
          ],
          "operationId": "saveBooking",
          "parameters": [
            {
              "name": "roomId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/register": {
        "post": {
          "tags": [
            "auth"
          ],
          "operationId": "register",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "tags": [
            "auth"
          ],
          "operationId": "login",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/LoginResponse"
                  }
                }
              }
            }
          }
        }
      },
...
      "/rooms/room-by-id/{roomId}": {
        "get": {
          "tags": [
            "room"
          ],
          "operationId": "getRoomById",
          "parameters": [
            {
              "name": "roomId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RoomDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/rooms/available-rooms-by-date-and-type": {
        "get": {
          "tags": [
            "room"
          ],
          "operationId": "getAllRoomsByDateAndType",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            },
            {
              "name": "checkInDate",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string",
                "format": "date"
              }
            },
            {
              "name": "checkOutDate",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string",
                "format": "date"
              }
            },
            {
              "name": "roomType",
              "in": "query",
              "required": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponseRoomDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/rooms/all": {
        "get": {
          "tags": [
            "room"
          ],
          "operationId": "getAllRooms",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponseRoomDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/rooms/all-available-rooms": {
        "get": {
          "tags": [
            "room"
          ],
          "operationId": "getAllAvailableRoom",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponseRoomDTO"
                  }
                }
              }
            }
          }
        }
      },
...
      "/booking/get-by-confirmation-code/{confirmationCode}": {
        "get": {
          "tags": [
            "Booking "
          ],
          "operationId": "findBookingByConfirmationCode",
          "parameters": [
            {
              "name": "confirmationCode",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookingDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/booking/all": {
        "get": {
          "tags": [
            "Booking "
          ],
          "operationId": "getAllBookings",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponseBookingDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/activate-account": {
        "get": {
          "tags": [
            "auth"
          ],
          "operationId": "activateAccount",
          "parameters": [
            {
              "name": "token",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/rooms/deleteRoom/{roomId}": {
        "delete": {
          "tags": [
            "room"
          ],
          "operationId": "deleteRoomById",
          "parameters": [
            {
              "name": "roomId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/booking/cancel/{bookingId}": {
        "delete": {
          "tags": [
            "Booking "
          ],
          "operationId": "cancelBooking",
          "parameters": [
            {
              "name": "bookingId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "RoomRegistrationRequest": {
          "required": [
            "roomDescription",
            "roomPrice",
            "roomType"
          ],
          "type": "object",
          "properties": {
            "roomType": {
              "type": "string"
            },
            "roomPrice": {
              "type": "number",
              "format": "double"
            },
            "roomDescription": {
              "type": "string"
            }
          }
        },
        "BookingDTO": {
          "type": "object",
          "properties": {
            "checkInDate": {
              "type": "string",
              "format": "date"
            },
            "checkOutDate": {
              "type": "string",
              "format": "date"
            },
            "numOfAdults": {
              "type": "integer",
              "format": "int32"
            },
            "numOfChildren": {
              "type": "integer",
              "format": "int32"
            },
            "totalNumOfGuest": {
              "type": "integer",
              "format": "int32"
            },
            "bookingConfigurationCode": {
              "type": "string"
            },
            "user": {
              "$ref": "#/components/schemas/UserDTO"
            },
            "room": {
              "$ref": "#/components/schemas/RoomDTO"
            },
            "id": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "Role": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string"
            },
            "createdDate": {
              "type": "string",
              "format": "date-time"
            },
            "lastModifiedDate": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "RoomDTO": {
          "required": [
            "roomDescription",
            "roomPrice",
            "roomType"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "roomType": {
              "type": "string"
            },
            "roomPhoto": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "byte"
              }
            },
            "roomDescription": {
              "type": "string"
            },
            "roomPrice": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "UserDTO": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "email": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "role": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Role"
              }
            },
            "phoneNumber": {
              "type": "string"
            },
            "bookings": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/BookingDTO"
              }
            }
          }
        },
        "RegisterRequest": {
          "required": [
            "email",
            "firstName",
            "lastName",
            "phoneNumber"
          ],
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "phoneNumber": {
              "type": "string"
            },
            "password": {
              "maxLength": 15,
              "minLength": 8,
              "type": "string"
            },
            "dateOfBirth": {
              "type": "string",
              "format": "date"
            },
            "email": {
              "type": "string"
            }
          }
        },
        "LoginRequest": {
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "maxLength": 15,
              "minLength": 8,
              "type": "string"
            }
          }
        },
        "LoginResponse": {
          "type": "object",
          "properties": {
            "token": {
              "type": "string"
            }
          }
        },
        "PageResponseBookingDTO": {
          "type": "object",
          "properties": {
            "content": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/BookingDTO"
              }
            },
            "number": {
              "type": "integer",
              "format": "int32"
            },
            "totalElements": {
              "type": "integer",
              "format": "int64"
            },
            "size": {
              "type": "integer",
              "format": "int32"
            },
            "totalPages": {
              "type": "integer",
              "format": "int32"
            },
            "first": {
              "type": "boolean"
            },
            "last": {
              "type": "boolean"
            }
          }
        },
...
        "PageResponseRoomDTO": {
          "type": "object",
          "properties": {
            "content": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/RoomDTO"
              }
            },
            "number": {
              "type": "integer",
              "format": "int32"
            },
            "totalElements": {
              "type": "integer",
              "format": "int64"
            },
            "size": {
              "type": "integer",
              "format": "int32"
            },
            "totalPages": {
              "type": "integer",
              "format": "int32"
            },
            "first": {
              "type": "boolean"
            },
            "last": {
              "type": "boolean"
            }
          }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "description": "JWT authentication description",
          "in": "header",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }